fix(coil-extension): always inject WM2, filter from background, closes #3074
Previously the content script would conditionally include the WM2 polyfills. It
did this by setting some state in the document.head.dataset before the <script>
injection then checking that state inside the polyfill code.

In Chrome and Firefox this seemed to work fine, but when testing Safari it was
discovered that there was a race condition between page executed code (which
referenced polyfill objects) and the injection.

In fact, by adding some logging of the relevant events, it was discovered that
the document readyState became `interactive` before an answer from the
background page could be received from the content script.

This means that there is no way in which a content script could conditionally
inject WM2 polyfills based upon dynamic information, only from build time
encoded data such as a static allowed list.

That doesn't mean that an origin trial is not possible, only that we'd have to
"live with" WM2 polyfill pollution where it's not needed. A bit ugly but not a
 showstopper. The background page orchestrator could still filter monetization
 requests based upon dynamic information.

This commit also adds a WEXT_BUILD_CONFIG_WM2_ALWAYS flag, which if defined
will bypass origin trial filtering.